"Rename": config 패키지 생성"
