+++ src/main/java/leggo/feed/backend/global/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package leggo.feed.backend.global;
+package leggo.feed.backend.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
