+++ src/main/java/leggo/feed/backend/global/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package leggo.feed.backend.global;
+package leggo.feed.backend.global.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
