+++ src/main/java/leggo/feed/backend/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package leggo.feed.backend.global;
+package leggo.feed.backend.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
